Update doc scripts to create download zip
