+++ packages/design-system-scripts/gulp/docs/docs.js
@@ -3,39 +3,13 @@
  * like parsing our CSS/JSX comments, generating JSON data files, and ultimately
  * generating the HTML files which get published as the public site.
  */
-const copyAssets = require('../common/copyAssets');
 const cleanDist = require('../common/cleanDist');
 const generatePages = require('./generatePages');
-const path = require('path');
+const { copySourceAssets, copyDocsAssets, copyDownloadZip } = require('./assets');
 const { compileDocsSass } = require('../sass');
 const { extractReactData } = require('./extractReactData');
-const { getDocsDirs } = require('../common/getDirsToProcess');
 const { logTask, log } = require('../common/logUtil');
 const { runWebpackStatically } = require('./webpack');
-
-/**
- * Copies all the fonts and images from the source package and the core design system package
- * In the case of a child DS, the source dir will already contain assets from the core npm package
- * from the `buildSrc` task that preceded `buildDocs`
- */
-function copySourceAssets(sourceDir, docsDir) {
-  logTask('🏞  ', `Copying fonts and images from source package into ${path.join(docsDir, 'dist')}`);
-  // Handle rootPath when copying
-  return copyAssets(path.join(sourceDir, 'dist'), path.join(docsDir, 'dist'));
-}
-
-/**
- * Copies all the fonts and images from our docs packages
- * Usually there will only be images in the docs package
- */
-async function copyDocsAssets(docsDir) {
-  logTask('🏞  ', `Copying fonts and images from docs packages into ${path.join(docsDir, 'dist')}`);
-  // Handle rootPath when copying
-  const docs = await getDocsDirs(docsDir);
-  return Promise.all([
-    docs.map((doc) => copyAssets(path.join(doc, 'src'), path.join(docsDir, 'dist'))),
-  ]);
-}
 
 module.exports = {
   /**
@@ -54,6 +28,7 @@ module.exports = {
     await cleanDist(docsDir);
     await copySourceAssets(sourceDir, docsDir);
     await copyDocsAssets(docsDir);
+    await copyDownloadZip(sourceDir, docsDir);
     await extractReactData(sourceDir, docsDir, options);
     await generatePages(sourceDir, docsDir, options);
     await compileDocsSass(docsDir, options);
